requirements.txt
Django==3.2.4
djangorestframework==3.12.4
redis==3.5.3
fakeredis==1.9.0